# src/main/java/com/liquibaseproject/api/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.liquibaseproject.api;

import com.liquibaseproject.model.ModelApiResponse;
import com.liquibaseproject.model.NewUser;
import java.util.UUID;
import com.liquibaseproject.model.Users;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "users", description = "List of users")
public interface UsersApi {

    default UsersApiDelegate getDelegate() {
        return new UsersApiDelegate() {};
    }

    /**
     * POST /users : Add a new user
     * Add a new user
     *
     * @param newUser Create a new user (required)
     * @return User has been successfully added (status code 200)
     *         or Access forbidden (status code 403)
     */
    @Operation(
        operationId = "addUser",
        summary = "Add a new user",
        description = "Add a new user",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User has been successfully added", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Users.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Users.class))
            }),
            @ApiResponse(responseCode = "403", description = "Access forbidden", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = ModelApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users",
        produces = { "application/json", "application/xml" },
        consumes = { "application/json", "application/xml", "application/x-www-form-urlencoded" }
    )
    
    default ResponseEntity<Users> addUser(
        @Parameter(name = "NewUser", description = "Create a new user", required = true) @Valid @RequestBody NewUser newUser
    ) {
        return getDelegate().addUser(newUser);
    }


    /**
     * DELETE /users/{id} : Deletes a user
     * delete a user
     *
     * @param id User ID for the user that needs to be deleted (required)
     * @return User found by ID (status code 200)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "Deletes a user",
        description = "delete a user",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User found by ID", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = ModelApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/users/{id}",
        produces = { "application/json", "application/xml" }
    )
    
    default ResponseEntity<ModelApiResponse> deleteUser(
        @Parameter(name = "id", description = "User ID for the user that needs to be deleted", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        return getDelegate().deleteUser(id);
    }


    /**
     * GET /users/findAll : List all users
     *
     * @return Users have been successfully listed (status code 200)
     */
    @Operation(
        operationId = "findAllUsers",
        summary = "List all users",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Users have been successfully listed", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Users.class))),
                @Content(mediaType = "application/xml", array = @ArraySchema(schema = @Schema(implementation = Users.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/findAll",
        produces = { "application/json", "application/xml" }
    )
    
    default ResponseEntity<List<Users>> findAllUsers(
        
    ) {
        return getDelegate().findAllUsers();
    }


    /**
     * GET /users/findByUsername : Find users by usernames
     * Multiple usernames can be provided with comma separated strings (case insensitive)
     *
     * @param usernames  (optional)
     * @return Users have been successfully found by username (status code 200)
     */
    @Operation(
        operationId = "findUsersByUsername",
        summary = "Find users by usernames",
        description = "Multiple usernames can be provided with comma separated strings (case insensitive)",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Users have been successfully found by username", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Users.class))),
                @Content(mediaType = "application/xml", array = @ArraySchema(schema = @Schema(implementation = Users.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/findByUsername",
        produces = { "application/json", "application/xml" }
    )
    
    default ResponseEntity<List<Users>> findUsersByUsername(
        @Parameter(name = "usernames", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "usernames", required = false) String usernames
    ) {
        return getDelegate().findUsersByUsername(usernames);
    }


    /**
     * GET /users/{id} : Find user by ID
     * Returns a single user
     *
     * @param id User ID to return (required)
     * @return User found by ID (status code 200)
     */
    @Operation(
        operationId = "getUserById",
        summary = "Find user by ID",
        description = "Returns a single user",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User found by ID", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Users.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Users.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/{id}",
        produces = { "application/json", "application/xml" }
    )
    
    default ResponseEntity<Users> getUserById(
        @Parameter(name = "id", description = "User ID to return", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        return getDelegate().getUserById(id);
    }


    /**
     * PUT /users : Update an existing user
     * Update an existing user by Id
     *
     * @param users Update an existing user (required)
     * @return User has been successfully updated (status code 200)
     *         or Access forbidden (status code 403)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Update an existing user",
        description = "Update an existing user by Id",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User has been successfully updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = ModelApiResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Access forbidden", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = ModelApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/users",
        produces = { "application/json", "application/xml" },
        consumes = { "application/json", "application/xml", "application/x-www-form-urlencoded" }
    )
    
    default ResponseEntity<ModelApiResponse> updateUser(
        @Parameter(name = "Users", description = "Update an existing user", required = true) @Valid @RequestBody Users users
    ) {
        return getDelegate().updateUser(users);
    }


    /**
     * POST /users/{id} : Updates a user with form data
     * 
     *
     * @param id ID of user that needs to be updated (required)
     * @param username Username for the user that needs to be updated (optional)
     * @param email E-mail of the user that needs to be updated (optional)
     * @return User found by ID (status code 200)
     *         or Access forbidden (status code 403)
     */
    @Operation(
        operationId = "updateUserWithForm",
        summary = "Updates a user with form data",
        description = "",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User found by ID", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = ModelApiResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Access forbidden", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = ModelApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users/{id}",
        produces = { "application/json", "application/xml" }
    )
    
    default ResponseEntity<ModelApiResponse> updateUserWithForm(
        @Parameter(name = "id", description = "ID of user that needs to be updated", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "username", description = "Username for the user that needs to be updated", in = ParameterIn.QUERY) @Valid @RequestParam(value = "username", required = false) String username,
        @Parameter(name = "email", description = "E-mail of the user that needs to be updated", in = ParameterIn.QUERY) @Valid @RequestParam(value = "email", required = false) String email
    ) {
        return getDelegate().updateUserWithForm(id, username, email);
    }

}