# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Liquibase Project - OpenAPI 3.0
  description: |-
    API for managing Users, Products, and Orders
  contact:
    email: <a href="mailto:stefan.golubov@qinshift.com">stefan.golubov@qinshift.com</a>
  version: 1.0.0
...
paths:
  /users:
    put:
      tags:
        - users
      summary: Update an existing user
      description: Update an existing user by Id
      operationId: updateUser
      requestBody:
        description: Update an existing user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
          application/xml:
            schema:
              $ref: '#/components/schemas/Users'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Users'
        required: true
      responses:
        '200':
          description: User has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/Users'
        '403':
          description: Access forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                type: Forbidden
                message: You don't have access to perform this operation
      security:
        - basicAuth: [ ]
    post:
      tags:
        - users
      summary: Add a new user
      description: Add a new user
      operationId: addUser
      requestBody:
        description: Create a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewUser'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewUser'
        required: true
      responses:
        '200':
          description: User has been successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
            application/xml:
              schema:
                $ref: '#/components/schemas/Users'
        '403':
          description: Access forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                type: Forbidden
                message: You don't have access to perform this operation
      security:
        - basicAuth: [ ]
  /users/findByUsername:
    get:
      tags:
        - users
      summary: Find users by usernames
      description: Multiple usernames can be provided with comma separated strings (case insensitive)
      operationId: findUsersByUsername
      parameters:
        - name: usernames
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Users have been successfully found by username
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Users'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Users'
      security:
        - basicAuth: [ ]
  /users/findAll:
    get:
      tags:
        - users
      summary: List all users
      operationId: findAllUsers
      responses:
        '200':
          description: Users have been successfully listed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Users'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Users'
      security:
        - basicAuth: [ ]
  /users/{id}:
    get:
      tags:
        - users
      summary: Find user by ID
      description: Returns a single user
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: User ID to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
            application/xml:
              schema:
                $ref: '#/components/schemas/Users'
      security:
        - basicAuth: [ ]
    post:
      tags:
        - users
      summary: Updates a user with form data
      description: ''
      operationId: updateUserWithForm
      parameters:
        - name: id
          in: path
          description: ID of user that needs to be updated
          required: true
          schema:
            type: string
            format: uuid
        - name: username
          in: query
          description: Username for the user that needs to be updated
          schema:
            type: string
        - name: email
          in: query
          description: E-mail of the user that needs to be updated
          schema:
            type: string
      responses:
        '200':
          description: User found by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Access forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                type: Forbidden
                message: You don't have access to perform this operation
      security:
        - basicAuth: [ ]
    delete:
      tags:
        - users
      summary: Deletes a user
      description: delete a user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: User ID for the user that needs to be deleted
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - basicAuth: [ ]
...
components:
  schemas:
    Users:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 6574214f-89e3-4001-b54d-0d185163f7a2
        username:
          type: string
          example: user.name
        email:
          type: string
          example: user@username.com
      xml:
        name: user
    NewUser:
      type: object
      properties:
        username:
          type: string
          example: user.name
        password:
          type: string
          example: A1&d_2ddr$%$
        email:
          type: string
          example: user@username.com
      xml:
        name: user
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: '##default'
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic